+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Feat: Spring WebClient integration (#1621)
+* Feat: OpenFeign integration (#1632)
 
 ## 5.1.0-beta.9